Updating API to reflect i32s

+++ sample-apps/discount-functions-sample-app/extensions/fixed-amount-product-discount/src/api.rs
@@ -2,7 +2,7 @@
 
 pub type Boolean = bool;
 pub type Float = f64;
-pub type Int = i64;
+pub type Int = i32;
 pub type ID = String;
 pub type Decimal = String;
